[FRONT][LIAISON][THEMES] Liaison back et front Bug sur bouton s'abonner / se désabonner...

+++ back/src/main/java/com/openclassrooms/mddapi/services/AbonnementService.java
@@ -2,7 +2,6 @@ package com.openclassrooms.mddapi.services;
 
 import com.openclassrooms.mddapi.dto.AbonnementRequestDTO;
 import com.openclassrooms.mddapi.dto.AbonnementResponseDTO;
-import com.openclassrooms.mddapi.dto.ArticleResponseDTO;
 import com.openclassrooms.mddapi.models.AbonnementModel;
 import com.openclassrooms.mddapi.models.UserModel;
 import com.openclassrooms.mddapi.repositories.AbonnementRepository;
@@ -12,6 +11,8 @@ import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
+
+import java.util.Optional;
 
 @Service
 public class AbonnementService implements IAbonnementService {
@@ -37,14 +38,21 @@ public class AbonnementService implements IAbonnementService {
     }
 
     @Override
-    public boolean unsubscribeTheme(Integer id, String userEmail) {
+    public boolean toggleSubscription(Integer id, String userEmail) {
         UserModel user = userRepository.findByEmail(userEmail)
                 .orElseThrow(() -> new UsernameNotFoundException("User not found with email"));
-        AbonnementModel abonnementModel = abonnementRepository.findByUserIdAndThemeId(user.getId(), id)
-                .orElseThrow(() ->
-                        new RuntimeException("Subscription not found for user with email: " + userEmail + " and theme ID: " + id));
+        Optional<AbonnementModel> abonnementModel = abonnementRepository.findByUserIdAndThemeId(user.getId(), id);
 
-        abonnementRepository.delete(abonnementModel);
+        if(abonnementModel.isPresent()){
+            abonnementRepository.delete(abonnementModel);
+        }
+        else{
+            AbonnementModel SubscriptionAbonnementModel = new AbonnementModel(id, user.getId());
+            AbonnementModel abonnementModelSaved =  abonnementRepository.save(SubscriptionAbonnementModel);
+            return true;
+        }
+
+
         return true;
     }
 }
